extend generator and system for comparable variants of each
