+++ src/interfaces/System.ts
@@ -1,0 +1,8 @@
+import { SystemState } from "../SystemState.ts";
+
+export interface System {
+    time: number;
+    state: Readonly<SystemState>;
+    setStateInitial: () => void;
+    runToTime: (time: number) => void;
+}
