+++ tests/DiscreteSystem.test.ts
@@ -1,37 +1,16 @@
-import { describe, test, expect, vi, afterEach, Mocked } from "vitest";
-import { discreteSIR, SIRData } from "./examples/discreteSIR";
+import { describe, test, expect, vi, afterEach } from "vitest";
+import { discreteSIR } from "./examples/discreteSIR";
 import { discreteWalk, WalkShared } from "./examples/discreteWalk.ts";
 import { DiscreteSystem } from "../src/DiscreteSystem";
 import { particleStateToArray } from "../src/utils";
 import { SIRShared } from "./examples/discreteSIR.ts";
 import { Random, RngStateBuiltin, RngStateObserved } from "@reside-ic/random";
-import { poissonLogDensity } from "../src/density.ts";
-import { SystemDataComparison } from "../src/SystemDataComparison.ts";
+import { expectedGroup1Initial, expectedGroup2Initial, shared } from "./examples/SIRTestHelpers.ts";
 
 const generator = discreteSIR;
-const shared = [
-    { N: 1000000, I0: 1, beta: 4, gamma: 2 },
-    { N: 2000000, I0: 2, beta: 8, gamma: 4 }
-];
-
-const expectedGroup1Initial = [
-    999999, // shared.N - shared.I0;
-    1, // shared.I0;
-    0,
-    0,
-    0
-];
-
-const expectedGroup2Initial = [
-    1999998, // shared.N - shared.I0;
-    2, // shared.I0;
-    0,
-    0,
-    0
-];
 
 const createSystem = (random?: Random) =>
-    new DiscreteSystem<SIRShared, null, SIRData>(
+    new DiscreteSystem<SIRShared, null>(
         generator,
         shared,
         5, // time
@@ -72,7 +51,7 @@ describe("DiscreteSystem", () => {
     });
 
     test("defaults to built in random", () => {
-        const sys = new DiscreteSystem<SIRShared, null, SIRData>(
+        const sys = new DiscreteSystem<SIRShared, null>(
             generator,
             shared,
             5, // time
@@ -86,7 +65,7 @@ describe("DiscreteSystem", () => {
     test("constructor throws error if nParticles is invalid", () => {
         expect(
             () =>
-                new DiscreteSystem<SIRShared, null, SIRData>(
+                new DiscreteSystem<SIRShared, null>(
                     generator,
                     shared,
                     5, // time
@@ -97,7 +76,7 @@ describe("DiscreteSystem", () => {
 
         expect(
             () =>
-                new DiscreteSystem<SIRShared, null, SIRData>(
+                new DiscreteSystem<SIRShared, null>(
                     generator,
                     shared,
                     5, // time
@@ -108,7 +87,7 @@ describe("DiscreteSystem", () => {
     });
 
     const expectParticleGroupState = (
-        sys: DiscreteSystem<any, any, any>,
+        sys: DiscreteSystem<any, any>,
         iGroup: number,
         nParticles: number,
         expectedValues: number[]
@@ -139,7 +118,7 @@ describe("DiscreteSystem", () => {
         const compareRandom = new Random(rngStateObserved.replay());
         const np = 5;
         const walkShared = { n: 3, sd: 1 };
-        const sys = new DiscreteSystem<WalkShared, null, null>(
+        const sys = new DiscreteSystem<WalkShared, null>(
             discreteWalk,
             [walkShared],
             0, // time
@@ -167,7 +146,7 @@ describe("DiscreteSystem", () => {
         const step1 = 5.5;
         const dt = 0.5;
 
-        const sys = new DiscreteSystem<SIRShared, null, SIRData>(
+        const sys = new DiscreteSystem<SIRShared, null>(
             generator,
             shared,
             start, // time
@@ -221,86 +200,5 @@ describe("DiscreteSystem", () => {
         expect(() => sys.runToTime(1)).toThrowError(
             "Cannot run to requested time 1, which is less than current time 5."
         );
-    });
-
-    test("can compare data", () => {
-        const genCompareDataSpy = vi.spyOn(generator, "compareData");
-        const sys = createSystem();
-        sys.setStateInitial(); // compare data with initial state where grp1 I = 1, and grp2 I = 2
-        const data = [{ prevalence: 2 }, { prevalence: 3 }];
-        const result = sys.compareData(data);
-        expect(result.nGroups).toBe(2);
-        expect(result.nParticles).toBe(3);
-        const expectedGrp1Value = poissonLogDensity(2, 1);
-        expect(result.getValue(0, 0)).toBe(expectedGrp1Value);
-        expect(result.getValue(0, 1)).toBe(expectedGrp1Value);
-        expect(result.getValue(0, 2)).toBe(expectedGrp1Value);
-        const expectedGrp2Value = poissonLogDensity(3, 2);
-        expect(result.getValue(1, 0)).toBe(expectedGrp2Value);
-        expect(result.getValue(1, 1)).toBe(expectedGrp2Value);
-        expect(result.getValue(1, 2)).toBe(expectedGrp2Value);
-
-        expect(genCompareDataSpy).toHaveBeenCalledTimes(6);
-        const expectedGrp1Params = [5, expectedGroup1Initial, data[0], shared[0], null, sys["_random"]];
-        expect(genCompareDataSpy.mock.calls[0]).toStrictEqual(expectedGrp1Params);
-        expect(genCompareDataSpy.mock.calls[1]).toStrictEqual(expectedGrp1Params);
-        expect(genCompareDataSpy.mock.calls[2]).toStrictEqual(expectedGrp1Params);
-        const expectedGrp2Params = [5, expectedGroup2Initial, data[1], shared[1], null, sys["_random"]];
-        expect(genCompareDataSpy.mock.calls[3]).toStrictEqual(expectedGrp2Params);
-        expect(genCompareDataSpy.mock.calls[4]).toStrictEqual(expectedGrp2Params);
-        expect(genCompareDataSpy.mock.calls[5]).toStrictEqual(expectedGrp2Params);
-    });
-
-    const expectSharedDataResult = (
-        result: SystemDataComparison,
-        data: SIRData,
-        genCompareDataSpy: Mocked<any>,
-        random: Random
-    ) => {
-        expect(result.nGroups).toBe(2);
-        expect(result.nParticles).toBe(3);
-        const expectedGrp1Value = poissonLogDensity(2, 1);
-        expect(result.getValue(0, 0)).toBe(expectedGrp1Value);
-        expect(result.getValue(0, 1)).toBe(expectedGrp1Value);
-        expect(result.getValue(0, 2)).toBe(expectedGrp1Value);
-        const expectedGrp2Value = poissonLogDensity(2, 2);
-        expect(result.getValue(1, 0)).toBe(expectedGrp2Value);
-        expect(result.getValue(1, 1)).toBe(expectedGrp2Value);
-        expect(result.getValue(1, 2)).toBe(expectedGrp2Value);
-
-        expect(genCompareDataSpy).toHaveBeenCalledTimes(6);
-        const expectedGrp1Params = [5, expectedGroup1Initial, data, shared[0], null, random];
-        expect(genCompareDataSpy.mock.calls[0]).toStrictEqual(expectedGrp1Params);
-        expect(genCompareDataSpy.mock.calls[1]).toStrictEqual(expectedGrp1Params);
-        expect(genCompareDataSpy.mock.calls[2]).toStrictEqual(expectedGrp1Params);
-        const expectedGrp2Params = [5, expectedGroup2Initial, data, shared[1], null, random];
-        expect(genCompareDataSpy.mock.calls[3]).toStrictEqual(expectedGrp2Params);
-        expect(genCompareDataSpy.mock.calls[4]).toStrictEqual(expectedGrp2Params);
-        expect(genCompareDataSpy.mock.calls[5]).toStrictEqual(expectedGrp2Params);
-    };
-
-    test("can compare with shared data - single item array", () => {
-        const genCompareDataSpy = vi.spyOn(generator, "compareData");
-        const sys = createSystem();
-        sys.setStateInitial();
-        const data = [{ prevalence: 2 }];
-        const result = sys.compareData(data);
-        expectSharedDataResult(result, data[0], genCompareDataSpy, sys["_random"]);
-    });
-
-    test("can compare with shared data - object", () => {
-        const genCompareDataSpy = vi.spyOn(generator, "compareData");
-        const sys = createSystem();
-        sys.setStateInitial();
-        const data = { prevalence: 2 };
-        const result = sys.compareData(data);
-        expectSharedDataResult(result, data, genCompareDataSpy, sys["_random"]);
-    });
-
-    test("compareData throws error if data is unexpected length", () => {
-        const sys = createSystem();
-        sys.setStateInitial();
-        const data = [{ prevalence: 1 }, { prevalence: 2 }, { prevalence: 3 }];
-        expect(() => sys.compareData(data)).toThrowError("Expected data to have same length as groups.");
     });
 });
